+++ mailer/src/events/listeners/resetPasswordListener.ts
@@ -1,7 +1,8 @@
 import { Listener, Subjects, ResetPasswordEvent } from '@tcosmin/common';
 import { Message } from 'node-nats-streaming';
-import { queueGroupName } from './queue-group-name';
-import { transporter } from '../../services/transporter';
+import { queueGroupName } from './queueGroupName';
+import { MailerService } from '../../services/mailerService';
+import Container from 'typedi';
 
 export class PasswordResetListener extends Listener<ResetPasswordEvent> {
   subject: Subjects.ResetPassword = Subjects.ResetPassword;
@@ -11,7 +12,8 @@ export class PasswordResetListener extends Listener<ResetPasswordEvent> {
   async onMessage(data: ResetPasswordEvent['data'], msg: Message) {
     const { email, resetToken } = data;
 
-    await transporter.sendResetPasswordEmail(email, resetToken);
+    const mailerService = Container.get(MailerService);
+    await mailerService.sendResetPasswordEmail(email, resetToken);
 
     msg.ack();
   }
